requirements.txt
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0
langfuse>=3.0.0
prompt_toolkit>=3.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
